+++ code/src/cljs/sixsq/nuvla/ui/notifications/subs.cljs
@@ -43,6 +43,17 @@
   (fn [db]
     (::spec/notification-subscription-configs db)))
 
+(def res-kind->sort-order {"nuvlabox" 1
+                           "data-record" 2})
+
+(reg-sub
+  ::notification-subscription-configs-grouped
+  :<- [::notification-subscription-configs]
+  (fn [notification-subscription-configs]
+    (->> (sort-by #(or (res-kind->sort-order (:resource-kind %)) 99)
+                 notification-subscription-configs)
+        (group-by :resource-kind))))
+
 (reg-sub
   ::notification-subscription-config
   (fn [db]