+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - NuvlaEdge details - Fix warnings in console
 - NuvlaEdge details - Always shows tags card in nuvlaedge details overview, bugfix #1101
 - Notification subscriptions - Addmodal: Disables greying out of unchecked radio button
+- Notification subscriptions - Sort config always showing NuvlaEdge first
 
 ## [2.32.12] - 2023-01-06